# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * $MAIL — Solana wallet mailbox API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ApiMessage {
  message: string;
}

export type Folder = typeof Folder[keyof typeof Folder];


export const Folder = {
  mailbox: 'mailbox',
  requests: 'requests',
  junk: 'junk',
  sent: 'sent',
} as const;

/**
 * addressed: recorded for the recipient wallet. delivered: surfaced on a $MAIL surface the recipient accessed. opened: the authenticated recipient opened it.
 */
export type MessageStatus = typeof MessageStatus[keyof typeof MessageStatus];


export const MessageStatus = {
  addressed: 'addressed',
  delivered: 'delivered',
  opened: 'opened',
} as const;

export interface AuthChallengeInput {
  /** Base58 Solana wallet address */
  wallet: string;
}

export interface AuthChallenge {
  wallet: string;
  nonce: string;
  /** Human-readable message to sign. Signing proves mailbox ownership. No SOL moves, no transaction is submitted. */
  message: string;
}

export interface AuthVerifyInput {
  wallet: string;
  nonce: string;
  /** Base58-encoded ed25519 signature of the challenge message */
  signature: string;
}

export interface Session {
  /**
     * The signed-in wallet address, or null when signed out
     * @nullable
     */
  wallet: string | null;
  authenticated: boolean;
}

export interface MailboxSummary {
  wallet: string;
  /** True when unread mail exists in the Mailbox — drives the flag-up animation */
  hasNewMail: boolean;
  mailboxUnread: number;
  mailboxCount: number;
  requestsCount: number;
  junkCount: number;
  sentCount: number;
}

export interface Message {
  id: string;
  senderWallet: string;
  recipientWallet: string;
  body: string;
  folder: Folder;
  status: MessageStatus;
  /** True when the sender proved wallet ownership by signature at send time */
  senderVerified: boolean;
  containsUrl: boolean;
  /** Destination domains detected in the body, for link-safety warnings */
  urlDomains: string[];
  threadId: string;
  /** @nullable */
  replyToMessageId: string | null;
  createdAt: string;
  /** @nullable */
  deliveredAt: string | null;
  /** @nullable */
  openedAt: string | null;
  /** True when the signed-in wallet is the sender */
  isOutgoing: boolean;
}

export interface MessageInput {
  /** Any valid base58 Solana wallet address — recipient does not need to be a $MAIL user */
  recipientWallet: string;
  /**
     * @minLength 1
     * @maxLength 4000
     */
  body: string;
  /** Set when replying to keep correspondence in one thread */
  replyToMessageId?: string;
}

export interface ShareCard {
  messageId: string;
  /** e.g. "I GOT MAIL" */
  headline: string;
  /** e.g. "STANDARD MAIL", "SIGNED MAIL" */
  deliveryType: string;
  senderVerified: boolean;
  receivedAt: string;
  /** False by default. The message stays PRIVATE unless the owner explicitly reveals it. */
  revealed: boolean;
  /**
     * Message content — null unless explicitly revealed by the owner
     * @nullable
     */
  body: string | null;
}

export interface BlockInput {
  wallet: string;
}

export interface BlockedWallet {
  wallet: string;
  createdAt: string;
}

export interface ReportInput {
  messageId: string;
  /**
     * @minLength 1
     * @maxLength 500
     */
  reason: string;
}

export interface Report {
  id: string;
  messageId: string;
  reason: string;
  createdAt: string;
}

/**
 * Where mail from unknown senders lands
 */
export type PreferencesUnknownSenderHandling = typeof PreferencesUnknownSenderHandling[keyof typeof PreferencesUnknownSenderHandling];


export const PreferencesUnknownSenderHandling = {
  requests: 'requests',
  mailbox: 'mailbox',
  junk: 'junk',
} as const;

export interface Preferences {
  wallet: string;
  /** Where mail from unknown senders lands */
  unknownSenderHandling: PreferencesUnknownSenderHandling;
}

export type PreferencesUpdateUnknownSenderHandling = typeof PreferencesUpdateUnknownSenderHandling[keyof typeof PreferencesUpdateUnknownSenderHandling];


export const PreferencesUpdateUnknownSenderHandling = {
  requests: 'requests',
  mailbox: 'mailbox',
  junk: 'junk',
} as const;

export interface PreferencesUpdate {
  unknownSenderHandling?: PreferencesUpdateUnknownSenderHandling;
}

export type ListMessagesParams = {
folder: Folder;
};