# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: $MAIL — Solana wallet mailbox API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /auth/challenge:
    post:
      operationId: createAuthChallenge
      tags: [auth]
      summary: Get a sign-in challenge for a wallet
      description: Returns a nonce message the wallet must sign to prove ownership. No transaction, no SOL moves.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthChallengeInput"
      responses:
        "200":
          description: Challenge created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AuthChallenge"
        "400":
          description: Invalid wallet address
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
  /auth/verify:
    post:
      operationId: verifyAuthSignature
      tags: [auth]
      summary: Verify a signed challenge and start a session
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AuthVerifyInput"
      responses:
        "200":
          description: Session started
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Session"
        "401":
          description: Signature verification failed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApiMessage"
...
components:
  schemas:
...
    ApiMessage:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...
    AuthChallengeInput:
      type: object
      properties:
        wallet:
          type: string
          description: Base58 Solana wallet address
      required:
        - wallet
    AuthChallenge:
      type: object
      properties:
        wallet:
          type: string
        nonce:
          type: string
        message:
          type: string
          description: Human-readable message to sign. Signing proves mailbox ownership. No SOL moves, no transaction is submitted.
      required:
        - wallet
        - nonce
        - message
    AuthVerifyInput:
      type: object
      properties:
        wallet:
          type: string
        nonce:
          type: string
        signature:
          type: string
          description: Base58-encoded ed25519 signature of the challenge message
      required:
        - wallet
        - nonce
        - signature
    Session:
      type: object
      properties:
        wallet:
          type: ["string", "null"]
          description: The signed-in wallet address, or null when signed out
        authenticated:
          type: boolean
      required:
        - wallet
        - authenticated
...