# lib/api-zod/src/generated/types/authChallengeInput.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Api
 * $MAIL — Solana wallet mailbox API
 * OpenAPI spec version: 0.1.0
 */

export interface AuthChallengeInput {
  /** Base58 Solana wallet address */
  wallet: string;
}